Added @Nullable and @NotNull. Refactoring

+++ src/main/java/ru/levin/tm/command/project/ProjectRemoveAllCommand.java
@@ -1,5 +1,7 @@
 package ru.levin.tm.command.project;
 
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import ru.levin.tm.api.IServiceLocator;
 import ru.levin.tm.api.service.IProjectService;
 import ru.levin.tm.api.service.ITaskService;
@@ -7,12 +9,19 @@ import ru.levin.tm.api.service.ITerminalService;
 import ru.levin.tm.command.AbstractCommand;
 
 public final class ProjectRemoveAllCommand extends AbstractCommand {
+    @NotNull
     private static final String ALL_PROJECTS_REMOVED_MESSAGE = "[ALL PROJECTS REMOVED]\n";
+
+    @NotNull
     private final IProjectService projectService;
+
+    @NotNull
     private final ITaskService taskService;
+
+    @NotNull
     private final ITerminalService terminalService;
 
-    public ProjectRemoveAllCommand(final IServiceLocator bootstrap) {
+    public ProjectRemoveAllCommand(@NotNull final IServiceLocator bootstrap) {
         super(bootstrap);
         this.projectService = bootstrap.getProjectService();
         this.taskService = bootstrap.getTaskService();
@@ -20,16 +29,19 @@ public final class ProjectRemoveAllCommand extends AbstractCommand {
     }
 
     @Override
+    @NotNull
     public String getName() {
         return "project-clear";
     }
 
     @Override
+    @NotNull
     public String getTitle() {
         return "";
     }
 
     @Override
+    @NotNull
     public String getDescription() {
         return "Remove all projects";
     }
@@ -41,11 +53,13 @@ public final class ProjectRemoveAllCommand extends AbstractCommand {
 
     @Override
     public void execute() {
-        final String userId = bootstrap.getUserService().getCurrentUser().getId();
+        if (bootstrap.getUserService().getCurrentUser() == null) return;
+        @Nullable final String userId = bootstrap.getUserService().getCurrentUser().getId();
+        if (userId == null) return;
         projectService.removeByUserId(userId);
 
         taskService.getAll().forEach(task -> {
-            if (task.getProjectId() != null && task.getUserId().equals(userId)) {
+            if (task.getProjectId() != null && userId.equals(task.getUserId())) {
                 taskService.remove(task);
             }
         });
